refactor: add_action_item to use bolt
